feat: merge render
