+++ packages/ice/template/entry.client.ts.ejs
@@ -1,6 +1,6 @@
-import { runApp } from '@ice/runtime';
+import { runBrowserApp } from '@ice/runtime';
 import appConfig from '@/app';
 import runtimeModules from './runtimeModules';
 import routes from './routes.client';
 
-runApp(appConfig, runtimeModules, routes);
+runBrowserApp(appConfig, runtimeModules, routes);
