+++ packages/runtime/src/index.ts
@@ -4,8 +4,7 @@ import {
 } from 'react-router-dom';
 import Runtime from './runtime.js';
 import App from './App.js';
-import render from './render.js';
-import runApp from './runApp.js';
+import runBrowserApp from './runBrowserApp.js';
 import runServerApp from './runServerApp.js';
 import { useAppContext } from './AppContext.js';
 import {
@@ -18,8 +17,7 @@ import {
 export {
   Runtime,
   App,
-  render,
-  runApp,
+  runBrowserApp,
   runServerApp,
   useAppContext,
   Link,
