+++ packages/runtime/src/runBrowserApp.tsx
@@ -1,12 +1,59 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { HashRouter, BrowserRouter, matchRoutes } from 'react-router-dom';
-import type Runtime from './runtime.js';
+import Runtime from './runtime.js';
 import App from './App.js';
 import AppRoutes from './AppRoutes.js';
-import type { AppRouterProps } from './types';
+import type { AppRouterProps, AppContext, InitialContext, AppConfig } from './types';
 
-export default async function render(runtime: Runtime) {
+export default async function runApp(config: AppConfig, runtimeModules, routes) {
+  const appConfig: AppConfig = {
+    ...config,
+    app: {
+      rootId: 'root',
+      strict: true,
+      ...(config?.app || {}),
+    },
+    router: {
+      type: 'browser',
+      ...(config?.router || {}),
+    },
+  };
+
+  const appContext: AppContext = {
+    routes,
+    appConfig,
+    initialData: null,
+  };
+
+  // ssr enabled and the server has returned data
+  if ((window as any).__ICE_APP_DATA__) {
+    appContext.initialData = (window as any).__ICE_APP_DATA__;
+    // context.pageInitialProps = (window as any).__ICE_PAGE_PROPS__;
+  } else if (appConfig?.app?.getInitialData) {
+    const { href, origin, pathname, search } = window.location;
+    const path = href.replace(origin, '');
+    // const query = queryString.parse(search);
+    const query = {};
+    const ssrError = (window as any).__ICE_SSR_ERROR__;
+    const initialContext: InitialContext = {
+      pathname,
+      path,
+      query,
+      ssrError,
+    };
+    appContext.initialData = await appConfig.app.getInitialData(initialContext);
+  }
+
+  const runtime = new Runtime(appContext);
+  runtimeModules.forEach(m => {
+    runtime.loadModule(m);
+  });
+
+  render(runtime);
+}
+
+async function render(runtime: Runtime) {
   const appContext = runtime.getAppContext();
   const { appConfig, routes } = appContext;
   const { rootId } = appConfig.app;
