+++ packages/runtime/src/runServerApp.tsx
@@ -1,6 +1,10 @@
+import * as React from 'react';
+import * as ReactDOMServer from 'react-dom/server.js';
+import { StaticRouter } from 'react-router-dom/server.js';
 import Runtime from './runtime.js';
-import serverRender from './serverRender.js';
-import type { AppContext, AppConfig } from './types';
+import App from './App.js';
+import AppRoutes from './AppRoutes.js';
+import type { AppContext, AppConfig, AppRouterProps } from './types';
 
 export default async function runServerApp(
     requestContext,
@@ -39,5 +43,39 @@ export default async function runServerApp(
     runtime.loadModule(m);
   });
 
-  return serverRender(runtime, requestContext, Document, documentOnly);
+  return render(runtime, requestContext, Document, documentOnly);
+}
+
+async function render(
+  runtime: Runtime,
+  requestContext,
+  Document,
+  documentOnly: boolean,
+) {
+  const documentHtml = ReactDOMServer.renderToString(<Document />);
+
+  if (documentOnly) {
+    return documentHtml;
+  }
+
+  let AppRouter = runtime.getAppRouter();
+  if (!AppRouter) {
+    const { req } = requestContext;
+    AppRouter = (props: AppRouterProps) => (
+      <StaticRouter location={req.url}>
+        <AppRoutes PageWrappers={props.PageWrappers} />
+      </StaticRouter>
+    );
+    runtime.setAppRouter(AppRouter);
+  }
+
+  const pageHtml = ReactDOMServer.renderToString(
+    <App
+      runtime={runtime}
+    />,
+  );
+
+  const html = documentHtml.replace('<!--app-html-->', pageHtml);
+
+  return html;
 }
